Arduino Code
Arduino code illustrating the evaluation of moisture percentage ranges based on our own table calculations from the digital signal given by the moisture sensors. The code also shows that it returns the given moisture level information in the form of a string for each sensor to the bluetooth module connected to the transmitter port of the Arduino. Then the strings are outputted on a bluetooth compatible app on a bluetooth compatible smartphone to read the moisture percentage levels of each of the different soils the sensors are occupying. This project was for my Senior design project at UCI and mentioned on my resume. Please watch our youtube video presentation at https://www.youtube.com/watch?v=iDwQdJlEDFs